zellij: adjust keybindings
Use Ctrl-V to enter locked mode and Shift-Escape to exit locked mode to
behieve the same as in qutebrowser.

Zellij does not support Shift as a modifier key, so use a unicode char
instead, which is U+011b, 01 stands for the shift key in CSI and 1b
stands for the escape key in ascii. The char is mapped in foot terminal
to Shift-Escape.